refactored server.ts to extract middleware and updated gitignore to prevent committing api.log
