+++ .gitignore
@@ -1,4 +1,5 @@
 .DS_Store
+api.log
 npm-debug.log
 node_modules
 .env
